alx_travel_app/requirement.txt
Django
djangorestframework
django-cors-headers
django-environ
drf-yasg
celery
mysqlclient
rabbitmq